refactor: :art: modularizing family' domain and infrastructure
created a new modular structure to better visualize boundaries of family information
