+++ family/domain/entity/const.go
