+++ family/infrastructure/repository/family/inMemory/findFamilyById.go
@@ -1,0 +1,20 @@
+package familyInMemory
+
+// import (
+// 	family "github.com/renatospaka/emr/family/domain/entity"
+// )
+
+// // Find a family by its ID
+// // or returns an error
+// func (f *FamilyRepositoryInMemory) FindById(id string) (*family.Family, error) {
+// 	if len(f.family) == 0 {
+// 		return &family.Family{}, family.ErrFamilyNotFound
+// 	}
+
+// 	for x, fam := range f.family {
+// 		if fam.ID == id {
+// 			return &f.family[x], nil
+// 		}
+// 	}
+// 	return &family.Family{}, family.ErrFamilyNotFound
+// }
