+++ family/infrastructure/repository/family/inMemory/remove_test.go
@@ -1,0 +1,15 @@
+package familyInMemory_test
+
+// import (
+// 	"testing"
+
+// 	"github.com/stretchr/testify/require"
+// )
+
+// func TestMember_Remove(t *testing.T) {
+// 	_ = repoMember.Add(member)
+	
+// 	id := member.ID
+// 	err := repoMember.Remove(id)
+// 	require.Nil(t, err)
+// }
